requirement.txt
annotated-types              0.7.0
filelock                     3.16.1
fsspec                       2024.9.0
generativeai                 0.0.1
google-ai-generativelanguage 0.6.10
google-api-core              2.21.0
google-api-python-client     2.149.0
google-auth                  2.35.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.65.0
grpcio                       1.67.0
grpcio-status                1.67.0
httplib2                     0.22.0
huggingface-hub              0.25.2
proto-plus                   1.24.0
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pydantic                     2.9.2
pydantic_core                2.23.4
pyparsing                    3.2.0
PyYAML                       6.0.2
rsa                          4.9
tqdm                         4.66.5
uritemplate                  4.1.1